% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisyphus.R
\name{sisyphus_run}
\alias{sisyphus_run}
\alias{sisyphus_stop}
\alias{sisyphus_change_delay}
\alias{sisyphus_change_check_fun}
\alias{sisyphus_change_fun_files_to_watch}
\alias{sisyphus_change_fun_files_to_ignore}
\alias{sisyphus_get_r_and_tests}
\alias{sisyphus_get_testthat_snaps}
\title{Run & Stop Sisyphus}
\usage{
sisyphus_run(
  check_fun = testthat::test_local,
  delay = 1,
  fun_files_to_watch = sisyphus::sisyphus_get_r_and_tests,
  fun_files_to_ignore = sisyphus::sisyphus_get_testthat_snaps
)

sisyphus_stop()

sisyphus_change_delay(delay)

sisyphus_change_check_fun(check_fun)

sisyphus_change_fun_files_to_watch(fun_files_to_watch)

sisyphus_change_fun_files_to_ignore(fun_files_to_ignore)

sisyphus_get_r_and_tests(home = getwd())

sisyphus_get_testthat_snaps(home = getwd())
}
\arguments{
\item{check_fun}{Function to run when a file is edited}

\item{delay}{Delay between checks}

\item{fun_files_to_watch}{Function that return a list of files to watch for changes}

\item{fun_files_to_ignore}{Function that return a list of files to ignore the changes}

\item{home}{Home directory from where to get R & tests files}
}
\value{
Used for side effect
}
\description{
Run & Stop Sisyphus
}
